Django-Middleware-0x03/0x03-MessagingApp-Django/requirements.txt
asgiref==3.8.1
Django==5.2.2
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
